requirements.txt
anyascii==0.3.3
asgiref==3.10.0
aws-requests-auth==0.4.3
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.7
django-cogwheels==0.3
django-debug-toolbar==6.0.0
django-extensions==4.1
django-filter==25.2
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.7
django-taggit==6.1.0
django-tasks==0.8.1
django-treebeard==4.7.1
djangorestframework==3.16.1
draftjs_exporter==5.1.0
elastic-transport==9.1.0
elasticsearch==9.1.1
et_xmlfile==2.0.0
filetype==1.2.0
idna==3.11
laces==0.1.2
openpyxl==3.1.5
pillow==11.3.0
pillow_heif==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wagtail==7.1.1
wagtail-font-awesome-svg==2.0
wagtailmenus==4.0.4
Willow==1.11.0
